Flush loglines atomically
Ruby's IO#puts calls write(2) twice: once to write the string, and once
to write a newline. This means that other writes to the same file (in
other processes, for instance) might be interleaved between einhorn log
lines. If lines explicitly end in the newline character, however, Ruby
will omit its newline and thus the second call to write(2).